requirements.txt
google-genai
python-dotenv
python-telegram-bot
requests
beautifulsoup4
markdownify
fastapi
uvicorn
websockets
